refactor(account-tokens): move token key semantics
