+++ src/services/accounts/utils/apiServiceRequest.ts
@@ -1,9 +1,9 @@
 import { shouldDecorateAccountApiRequestWithAuthSession } from "~/services/accounts/accountSiteProfile"
 import { accountSub2ApiAuthSession } from "~/services/accounts/sub2apiAuthSession"
+import { formatOptionalSkPrefixSiteToken } from "~/services/accountTokens/apiTokenKey"
 import type { KeyManagementCapability } from "~/services/apiAdapters/contracts/keyManagement"
 import type { TokenProvisioningCapability } from "~/services/apiAdapters/contracts/tokenProvisioning"
 import { getSiteAdapter } from "~/services/apiAdapters/registry"
-import { formatOptionalSkPrefixSiteToken } from "~/services/apiService/common/apiKey"
 import type { ApiServiceRequest } from "~/services/apiService/common/type"
 import type { Sub2ApiAuthSessionRequest } from "~/services/apiService/sub2api/authSession"
 import { AuthTypeEnum, type ApiToken, type DisplaySiteData } from "~/types"
